+++ src/basic/events/index.js
@@ -1,1 +1,2 @@
 export { setupManualEvents } from './manual-events.js';
+export { setupAddToCartEvent, setupCartItemEvents } from './cart-events.js';
